+++ packages/admin-color-picker/src/themeAugmentation.ts
@@ -1,7 +1,14 @@
-import { VPAdminColorPickerClassKeys } from "./core/ColorPicker.styles";
+import { ColorPickerThemeProps } from "./core/ColorPicker";
+import { CometAdminColorPickerClassKeys } from "./core/ColorPicker.styles";
+
+declare module "@material-ui/core/styles/props" {
+    interface ComponentsPropsList {
+        CometAdminColorPicker: ColorPickerThemeProps;
+    }
+}
 
 declare module "@material-ui/core/styles/overrides" {
     interface ComponentNameToClassKey {
-        VPAdminColorPicker: VPAdminColorPickerClassKeys;
+        CometAdminColorPicker: CometAdminColorPickerClassKeys;
     }
 }