+++ packages/admin-icons/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./generated";
+
+export { default as ThreeDotSaving } from "./ThreeDotSaving";
+export { default as CometColor } from "./CometColor";