+++ packages/admin-rte/src/core/Controls/FeaturesButtonGroup.sc.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const ButtonWrapper = styled.div`
-    margin-right: 1px;
-
-    :last-child {
-        margin-right: 0;
-    }
-`;