+++ packages/admin-stories/src/admin/table/TableFilter.tsx
@@ -1,6 +1,6 @@
+import { gql } from "@apollo/client";
 import { Field, FinalFormInput, Table, TableFilterFinalForm, TableQuery, useTableQuery, useTableQueryFilter } from "@comet/admin";
 import { storiesOf } from "@storybook/react";
-import gql from "graphql-tag";
 import * as qs from "qs";
 import * as React from "react";