+++ packages/admin-theme/src/cometAdminOverrides/clearInputButton.ts
@@ -1,0 +1,8 @@
+import { CometAdminClearInputButtonClassKeys } from "@comet/admin";
+import { StyleRules } from "@material-ui/styles/withStyles";
+
+export default (): StyleRules<{}, CometAdminClearInputButtonClassKeys> => ({
+    root: {},
+    disabled: {},
+    defaultIcon: {},
+});