+++ packages/admin-theme/src/cometAdminOverrides/masterLayout.ts
@@ -1,0 +1,10 @@
+import { CometAdminMasterLayoutClassKeys } from "@comet/admin";
+import { StyleRules } from "@material-ui/styles/withStyles";
+
+export default (): StyleRules<{}, CometAdminMasterLayoutClassKeys> => ({
+    root: {},
+    header: {},
+    toolbar: {},
+    contentWrapper: {},
+    mainContent: {},
+});