+++ packages/admin-theme/src/cometAdminOverrides/menu.ts
@@ -1,0 +1,10 @@
+import { CometAdminMenuClassKeys } from "@comet/admin";
+import { StyleRules } from "@material-ui/styles/withStyles";
+
+export default (): StyleRules<{}, CometAdminMenuClassKeys> => ({
+    drawer: {},
+    permanent: {},
+    temporary: {},
+    open: {},
+    closed: {},
+});