+++ packages/admin/src/error/errorboundary/ErrorBoundary.sc.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const ExceptionContainerDetails = styled.details`
+    white-space: pre-wrap;
+`;
+
+export const ExceptionContainerSummary = styled.summary`
+    cursor: pointer;
+    outline: none;
+`;