+++ packages/admin/src/index.ts
@@ -26,3 +26,7 @@ export * from "./Selection";
 export * from "./SelectionApi";
 export * from "./SelectionRoute";
 export * from "./Tabs";
+
+export * from "./error";
+
+export * from "./hooks";